refactor: wasm StretcherNode
