+++ packages/react-native-audio-api/src/web-core/AudioBuffer.web.ts
@@ -1,6 +1,5 @@
 import { IndexSizeError, NotSupportedError } from '../errors';
-import { AudioBufferLike } from '../types';
-import { AudioBufferOptions } from '../types';
+import { AudioBufferLike, AudioBufferOptions } from '../types';
 
 export default class AudioBuffer implements AudioBufferLike {
   readonly length: number;
